# planung/restapi/api/src/main/java/org/openapitools/api/WeckerConfigurationApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.1.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.ModelConfiguration;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2023-11-19T14:12:38.522563+01:00[Europe/Berlin]")
@Validated
@Tag(name = "Wecker Configuration", description = "Everything about the Configuration of your Alarm Clock")
public interface WeckerConfigurationApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /wecker/configurations/ : Add Configuration with ID
     * Add Configuration with ID
     *
     * @param modelConfiguration Add Configuration (optional)
     * @return Successful operation (status code 200)
     *         or Invalid ID supplied (status code 400)
     *         or Configuration not found (status code 404)
     *         or Validation exception (status code 405)
     */
    @Operation(
        operationId = "addConfiguration",
        summary = "Add Configuration with ID",
        description = "Add Configuration with ID",
        tags = { "Wecker Configuration" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ModelConfiguration.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid ID supplied"),
            @ApiResponse(responseCode = "404", description = "Configuration not found"),
            @ApiResponse(responseCode = "405", description = "Validation exception")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/wecker/configurations/",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<ModelConfiguration> addConfiguration(
        @Parameter(name = "ModelConfiguration", description = "Add Configuration") @Valid @RequestBody(required = false) ModelConfiguration modelConfiguration
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"active\" : true, \"days\" : [ { \"name\" : \"Donnerstag\", \"active\" : true, \"id\" : 123 }, { \"name\" : \"Donnerstag\", \"active\" : true, \"id\" : 123 } ], \"id\" : 123, \"maxDuration\" : 123, \"timestamp\" : 123 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /wecker/configurations/{configurationId} : Delete Configuration with ID
     * Delete Configuration with ID
     *
     * @param configurationId ID of Configuration (required)
     * @return Successful operation (status code 200)
     *         or Invalid ID supplied (status code 400)
     *         or Configuration not found (status code 404)
     *         or Validation exception (status code 405)
     */
    @Operation(
        operationId = "deleteConfiguration",
        summary = "Delete Configuration with ID",
        description = "Delete Configuration with ID",
        tags = { "Wecker Configuration" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ModelConfiguration.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid ID supplied"),
            @ApiResponse(responseCode = "404", description = "Configuration not found"),
            @ApiResponse(responseCode = "405", description = "Validation exception")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/wecker/configurations/{configurationId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<ModelConfiguration> deleteConfiguration(
        @Parameter(name = "configurationId", description = "ID of Configuration", required = true, in = ParameterIn.PATH) @PathVariable("configurationId") Long configurationId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"active\" : true, \"days\" : [ { \"name\" : \"Donnerstag\", \"active\" : true, \"id\" : 123 }, { \"name\" : \"Donnerstag\", \"active\" : true, \"id\" : 123 } ], \"id\" : 123, \"maxDuration\" : 123, \"timestamp\" : 123 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /wecker/configurations/ : Get all Wecker Configurations
     * Get all Wecker Configurations
     *
     * @return Successful operation (status code 200)
     */
    @Operation(
        operationId = "getAllConfigurations",
        summary = "Get all Wecker Configurations",
        description = "Get all Wecker Configurations",
        tags = { "Wecker Configuration" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = ModelConfiguration.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/wecker/configurations/",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<ModelConfiguration>> getAllConfigurations(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"active\" : true, \"days\" : [ { \"name\" : \"Donnerstag\", \"active\" : true, \"id\" : 123 }, { \"name\" : \"Donnerstag\", \"active\" : true, \"id\" : 123 } ], \"id\" : 123, \"maxDuration\" : 123, \"timestamp\" : 123 }, { \"active\" : true, \"days\" : [ { \"name\" : \"Donnerstag\", \"active\" : true, \"id\" : 123 }, { \"name\" : \"Donnerstag\", \"active\" : true, \"id\" : 123 } ], \"id\" : 123, \"maxDuration\" : 123, \"timestamp\" : 123 } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /wecker/configurations/{configurationId} : Get all Wecker Configurations
     * Get all Wecker Configurations
     *
     * @param configurationId ID of Configuration (required)
     * @return Successful operation (status code 200)
     *         or Invalid ID supplied (status code 400)
     *         or Configuration not found (status code 404)
     *         or Validation exception (status code 405)
     */
    @Operation(
        operationId = "getConfigurations",
        summary = "Get all Wecker Configurations",
        description = "Get all Wecker Configurations",
        tags = { "Wecker Configuration" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ModelConfiguration.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid ID supplied"),
            @ApiResponse(responseCode = "404", description = "Configuration not found"),
            @ApiResponse(responseCode = "405", description = "Validation exception")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/wecker/configurations/{configurationId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<ModelConfiguration> getConfigurations(
        @Parameter(name = "configurationId", description = "ID of Configuration", required = true, in = ParameterIn.PATH) @PathVariable("configurationId") Long configurationId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"active\" : true, \"days\" : [ { \"name\" : \"Donnerstag\", \"active\" : true, \"id\" : 123 }, { \"name\" : \"Donnerstag\", \"active\" : true, \"id\" : 123 } ], \"id\" : 123, \"maxDuration\" : 123, \"timestamp\" : 123 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /wecker/configurations/{configurationId} : Update Configuration with ID
     * Update Configuration with ID
     *
     * @param configurationId ID of Configuration (required)
     * @param modelConfiguration Update an existent Configuration (optional)
     * @return Successful operation (status code 200)
     *         or Invalid ID supplied (status code 400)
     *         or Configuration not found (status code 404)
     *         or Validation exception (status code 405)
     */
    @Operation(
        operationId = "updateConfiguration",
        summary = "Update Configuration with ID",
        description = "Update Configuration with ID",
        tags = { "Wecker Configuration" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ModelConfiguration.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid ID supplied"),
            @ApiResponse(responseCode = "404", description = "Configuration not found"),
            @ApiResponse(responseCode = "405", description = "Validation exception")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/wecker/configurations/{configurationId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<ModelConfiguration> updateConfiguration(
        @Parameter(name = "configurationId", description = "ID of Configuration", required = true, in = ParameterIn.PATH) @PathVariable("configurationId") Long configurationId,
        @Parameter(name = "ModelConfiguration", description = "Update an existent Configuration") @Valid @RequestBody(required = false) ModelConfiguration modelConfiguration
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"active\" : true, \"days\" : [ { \"name\" : \"Donnerstag\", \"active\" : true, \"id\" : 123 }, { \"name\" : \"Donnerstag\", \"active\" : true, \"id\" : 123 } ], \"id\" : 123, \"maxDuration\" : 123, \"timestamp\" : 123 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}