# planung/restapi/api/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: Dies ist eine Schnittstellenbeschreibung für die Wecker-REST-API
  termsOfService: http://swagger.io/terms/
  title: Wecker Rest API - OpenAPI 3.0
  version: 3.1.0
servers:
- url: localhost/api/v1
...
paths:
  /wecker/configurations/:
    get:
      description: Get all Wecker Configurations
      operationId: getAllConfigurations
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Configuration'
                type: array
          description: Successful operation
      summary: Get all Wecker Configurations
      tags:
      - Wecker Configuration
      x-accepts: application/json
      x-tags:
      - tag: Wecker Configuration
    post:
      description: Add Configuration with ID
      operationId: addConfiguration
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Configuration'
        description: Add Configuration
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Configuration'
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Configuration not found
        "405":
          description: Validation exception
      summary: Add Configuration with ID
      tags:
      - Wecker Configuration
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Wecker Configuration
  /wecker/configurations/{configurationId}:
    delete:
      description: Delete Configuration with ID
      operationId: deleteConfiguration
      parameters:
      - description: ID of Configuration
        explode: false
        in: path
        name: configurationId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Configuration'
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Configuration not found
        "405":
          description: Validation exception
      summary: Delete Configuration with ID
      tags:
      - Wecker Configuration
      x-accepts: application/json
      x-tags:
      - tag: Wecker Configuration
    get:
      description: Get all Wecker Configurations
      operationId: getConfigurations
      parameters:
      - description: ID of Configuration
        explode: false
        in: path
        name: configurationId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Configuration'
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Configuration not found
        "405":
          description: Validation exception
      summary: Get all Wecker Configurations
      tags:
      - Wecker Configuration
      x-accepts: application/json
      x-tags:
      - tag: Wecker Configuration
    put:
      description: Update Configuration with ID
      operationId: updateConfiguration
      parameters:
      - description: ID of Configuration
        explode: false
        in: path
        name: configurationId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Configuration'
        description: Update an existent Configuration
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Configuration'
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Configuration not found
        "405":
          description: Validation exception
      summary: Update Configuration with ID
      tags:
      - Wecker Configuration
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Wecker Configuration
...
components:
...
  schemas:
...
    Day:
      example:
        name: Donnerstag
        active: true
        id: 123
      properties:
        id:
          example: 123
          format: int64
          type: integer
        name:
          example: Donnerstag
          type: string
        active:
          example: true
          type: boolean
      type: object
    Configuration:
      example:
        active: true
        days:
        - name: Donnerstag
          active: true
          id: 123
        - name: Donnerstag
          active: true
          id: 123
        id: 123
        maxDuration: 123
        timestamp: 123
      properties:
        id:
          example: 123
          format: int64
          type: integer
        timestamp:
          example: 123
          format: int64
          type: integer
        maxDuration:
          example: 123
          format: int64
          type: integer
        active:
          example: true
          type: boolean
        days:
          items:
            $ref: '#/components/schemas/Day'
          type: array
      type: object
...